requirements.txt
torch
torchvision
matplotlib
gradio
pillow
numpy